Add reference request business rules
This adds the logic related to when an application form is considered to
be ready for an assessor to review after receiving a response to a
reference request. Specifically, we allow the assessor to start
reviewing if we've received more than 20 months, or if there are no more
references to request.