Ensure we send a boolean value as written statement confirmation
This field is not nullable, so we should make sure that we send false
instead of nil. With a form that's got a single checkbox, if the box
isn't checked then we get no form params back, so instead of using
permit we can try and fetch for the single value. If it's not checked
it'll come in as nil, which will now be stored as false.