Split assessment section show view
This splits up the summary component of this page in to separate views
which are rendered according to the assessment section. This should
help to keep this page maintainable.

I've also been able to remove a few methods from the view object which
were mostly just delegating to the models which we can pass in directly.