Ensure written_statement_confirmation is a boolean
This follows up 76c9de0fe6969082d2c9174f007a82ff18f8be84 with a more
explicit fix. When the user chooses to "save and continue", the
validation doesn't run which means it's possible for this value to be
nil or something which isn't a boolean. To prevent errors when writing
to the database, we can ensure the value is a boolean.