Add English language proficiency test document failure reason
This adds a new failure reason for teachers who are using the reduced
evidence journey and are able to upload their English language
proficiency test results in the form of a document. We need to allow
them to upload an alternative document if it's not legible or in a
format we can't open.