Allow declining after award pending verification
If the work references and other verification requests aren't acceptable
then we need to allow the assessors to decline the application after
this point.

We haven't built the flow which verifies the requests yet so we can't be
specific on the criteria, but I've left a comment for us to build later.