Refactor RequestableReviewForm
This changes the requestable review form to allow reviewing requestables
which are also locatable, meaning there needs to be an extra radio
button asking the assessor whether they have gone to review the
requestable before they can mark whether they are satisfied with it.

This logic is necessary for professional standing requests and
qualification requests, but not reference requests.