Set induction required when submitting application
We only do this for applications where we accept reduced evidence, as
this is the earliest point we can calculate the value as we trust the
work history provided by the applicant. I've moved it out of the award
flow as that happens to late to determine whether awarding is possible,
and it's a GET request so shouldn't really be performing any actions.