Refactor eligibility check status
This changes how we calculate the status of an eligibility check to
ensure that it doesn't allow skipping steps, even if a previous value is
set correctly. For example, if a question after choosing the country is
set (for whatever reason) but the country is not, then the status
should still be left at country.