Handle UncaughtThrowError from warden
ActionController::Live module changes the "process" method so that it runs inside a spawned thread.
The "process" method will also handle all filters (before/around/after action hooks)
Usually the authentication happens in a before action filter and if the user is not authenticated Devise will throw :warden
but since this is running in a spawn thread, the Warden middleware doesn't have the chance to catch this symbol and handle it properly.

See https://github.com/rosenfeld/devise-and-streaming

This commit adds some additional authentication methods in a module to
handle actions which could call send_blob_stream or other streaming
methods after a Devise authenticate_teacher|staff! before hook.
For the reasons explained above this won't handle an unauthenticated
user gracefully, hence the defensive code.