Refactor ActiveStorage Faraday::Error filter
This refactors the Sentry filter we use to avoid unnecessary
Faraday::Error exceptions arriving in Sentry related to timeout errors
with Azure. I've refactored this as I've noticed that we haven't had a
single Sentry error reported since this code was first put in
138f749436376ea37b0b28aea505b285ed8dbb9c. Although I can't see anything
in our logs, I'm worried it's preventing us from receiving other Sentry
errors.

I've refactored the code to rely more on the `hint` provided to use by
Sentry, and the `extra` which I've confirmed has the structure I'm
relying on:
https://github.com/getsentry/sentry-ruby/blob/bad263cebe621ab779d80338f0d6016b8ea3820f/sentry-rails/lib/sentry/rails/active_job.rb#L61-L70