Add DOCKER_IMAGE parameter to Makefile
Currently the way to configure this is to set the Terraform variable
`apply_qts_docker_image` directly, usually via the environment variable
`TF_VAR_apply_qts_docker_image`. This method works, but it's not very
user friendly as we're bypassing the Makefile and can't put in some
error handling. By adding this new parameter we can ensure the value is
set before continuing.