Catch timeout errors before rendering uploads
Sometimes we get timeout errors when downloading a file from Azure to
show it to the user, and there isn't much we can do to handle this case.
Rather than seeing the error in Sentry, this proposes that we simply
render the internal server error page right away. This is no different
to what users are seeing currently, but it avoids us having Sentry
issues which we can't solve.