Increase delay before fetching scan result
We are seeing a significant number of uploads where the scan result is stuck in the pending state, this presumably means we have attempted to fetch the scan result too quickly.
This commit increases the delay from 1 minute to 2 minutes to give malware scanning more time before attempting to fetch the result.