Allow filling in expired reference requests
This updated the queries to allow viewing the reference request page if
the application hasn't been awarded or declined, meaning that referees
can still fill in a reference after it's expired.

We sometimes want this if the applicant provided an incorrect email
address and we've re-sent the email, the expiration date may have
passed, but the assessors haven't declined the application.

This was the intended behaviour but wasn't working correctly.