Set format for error responses
Currently when we render the error response we use the format of the
request, and this means if a user tries to access (for example) a CSS
file at the wrong URL, it tries to render `errors/not_found.css` which
doesn't exist.

We can an error like this:

12:46:39 web.1    | Error during failsafe response: Missing template errors/not_found, application/not_found with {:locale=>[:en], :formats=>[:css], :variants=>[], :handlers=>[:raw, :erb, :html, :builder, :ruby, :md]}.
12:46:39 web.1    |
12:46:39 web.1    | Searched in:
12:46:39 web.1    |   * "/Users/thomasleese/Developer/dfe/apply-for-qualified-teacher-status/app/views"

I've noticed this seems to be the main reason why we get 500 errors, so
fixing this should make it easier to search for real errors in the logs.