Allow assessing sections before professional standing
We only need to limit the professional standing section from being
assessed before the professional standing has been received. This also
fixes a bug currently where countries with preliminary checks aren't
able to assess the qualifications section before receiving the
professional standing, but in the case of a "quick decline" we'd want
them to be able to assess the qualification section.