Add alternative registration number failure reason
This failure reason explicitly asked for the applicant to upload a
document as opposed to asking for a textual item of further information.
We need these two variations are there are times when we need to ask for
a document and times when we don't want to ask for a document.