Refactor smoke tests
This refactors the smoke tests to allow them to work for both PaaS and
AKS deployments. I've simplified the smoke test action itself to require
only what is needed to run the smoke tests, and then extracted the
credentials according to the logic of AKS or PaaS, automatically calling
the smoke tests after each deployment.