Don't expire old tokens on sign in
This is to better handle Outlook's "safe links" feature where it
navigates to the page before the user does, and so in certain situations
we end up treating the user as already signed in when they come to do
it.

A better fix might be to look at the user-agent and not record the sign
in if it's coming from Outlook, but I want to see if this change fixes
the immediate problem.