Content changes for ineligable country kick out page
If a user picks a country that is not eligible for QTS application, they will see the ineligible country message page. This makes reference to all countries being online by the end of 2023. This now seems unlikely so we need to change the content on this page.

Second sentence changes from:

However, the regulations are changing, which means you’ll be eligible to use this service by the end of 2023.

To:

We will update this page should there be any changes to eligibility for QTS.

Note:

The link should go to the ‘Changes to eligibility for QTS’ section of this page https://www.gov.uk/government/publications/awarding-qualified-teacher-status-to-overseas-teachers/a-fairer-approach-to-awarding-qts-to-overseas-teachers--2#changes-to-eligibility-for-qts

Remove: Remove the current link to ‘Learn more about the qualifications and experience you’ll need to reply for QTS'.