Improve preliminary check status
This ensures that the status is always set correctly, and improves the
readability of the code. It's necessary if an assessors reviews the
preliminary assessment section but then chooses not to decline
immediately, this will currently result in the status changing to
"assessment in progress" but I think we should leave it as "preliminary
check".