Allow backfilling waiting on preliminary checks
If an application is in the waiting on status because the teaching
authority provides the written statement, we were previously ignoring it
as it wouldn't be in the "submitted" state.

However, this is incorrect, as some applications will have two
pre-assessment tasks, one waiting on the profession standing, and one to
complete the preliminary check.