Ask application if they have a document
For optional documents will start by asking the applicant if they've got
the document, and then if they do, we require them to upload a file. If
they don't have a document then they can skip the document upload
section.

At the moment the content is hard coded for the Northern Ireland
induction document as that's the only optional document, but it can be
expanded in the future if necessary.