Don't send reminder emails for old applications
If a teacher has two draft application forms (which shouldn't happen,
but is possible in rare circumstances such as resubmitting forms), then
currently we send them two emails about an application being deleted.
However, teachers can only access their most recent application form, so
we don't need to send an email about a previous one as it's not
accessible anyway.

This will also fix an issue in Sentry where we were sending emails for
the wrong application form as we send from the perspective of a teacher
(which only has access to the most recent application form):

https://dfe-teacher-services.sentry.io/issues/4322438918/