Add restrictions on rolling back assessments
This adds a number of restrictions on when we can rollback assessments
to avoid invalid states. Specifically we delete any draft applications
which the applicant may have started since receiving a decline email to
ensure they can't submit two in applications in parallel. It also
prevents the rollback if the applicant has submitted a second
application after receiving a decline email as we should be assessing
that instead of rolling back.