Add missing target for checkboxes
When an error occurs in the form we get a link which links to the
relevant field with the error. Unfortunately this doesn't work for the
checkboxes because we're not using the field as intented, instead we're
using a fieldset but each checkbox as got its own name, so we never end
up with a suitable target. This commit adds an invisible div target
which can be used to link to.