Improve status filter
Currently when we filter on the boolean statuses, we're not filtering on
their associated overall status, meaning we can get, for example, a
declined application where we're waiting on the further information
still. The assessors don't need to see this application as its overall
status is declined, so we can remove it from the list.