Refactor secondary_education_teaching_qualification_required?
This removes the method and instead gets all users of this method to
access the value directly. The reason for this is that we don't store
the subject limited value on the application form (i.e. it can change if
a country is made subject limited after an application is created).