Consider qualification when calculating work history duration
This updates the calculation to (optionally) include the teaching
qualification, which is currently missing from the calculation.
Effectively, this will disregard any months of work history that are
prior to the qualification certificate date.

This is currently behind a flag as we want to do some analyse on the
number of applications which will be affected by this change, with the
intention to eventually change the calculation to use this in all cases.