Use teaching_qualification_part_of_degree
This changes the code to use the teaching_qualification_part_of_degree
column which has been added to the application form for capturing
whether the teaching qualification is part of the degree. We're moving
away from using the part_of_university_degree column on the
qualification as there is only one entry for this.