Reset Sidekiq retry options
This removes the custom Sidekiq retry options for the DQT TRN Request
jobs. The 14 retries limit means we were retrying for approximately 1
day, but often the DQT issues will not get resolved so quickly. By
leaving it to the default settings, the job will be retried for almost 3
weeks before being dropped.

https://github.com/sidekiq/sidekiq/wiki/Error-Handling