Improve footer content
This improves the content displayed in the footer as requested by our
content designer to clarify information for applicants when they're
signed in and correct the respond to time.

I've also updated some tests that relied on the existance of a single
`h1` when we now have more than one.