Set Sidekiq log level to warning
This sets the global log level configuration to warning if we're running
as the Sidekiq service. This allows us to reduce the number of generic
logs related to when runs are run (which we have access to anyway via
the dashboard) and instead focus on logging which might require
investigation.