Don't hide decline reasons if expired
If the further information request or LoPS has expired, in most cases
this leads to an automatic decline. However, it's possible to rollback
this decision and allow the assessor to continue with the assessment.
Currently if the assessor then goes on to decline the application
manually, these reasons will not be shown.