Ghana specific registration number form
The Registration number, also known as the
license number for Ghanaian specific applicants
is a 3 part alphanumeric which has a specific
format. To make it easier for applicants to not
make mistakes when entering this number, we
have decided to change the UI with additional
validations in place to guide users.

Note: As part of this work a new grouped input
had to be introduced with inspiration taken from
the GOV design systems date field. Since, this
is a custom group of inputs, there are no helpers
available and as a result of this, custom input fields with GOV design system classes have been used to replicate the desired design.