Ensure potential_duplicate flag is set if potential_duplicate from TRS is true and no TRN is included in the response
This is because, regardless of any duplicates being resolved on TRS after it has already been flagged as a duplicate, TRS will always return `potential_duplicate` as true. This is different to how the V2 endpoints behaved. The new behaviour is that whenever the response includes `trn`, then that means the potential duplicate has been resolved whether through merging with an existing record or being marked as a false negative.