Rename, remove and add columns related to SLA for application_forms, assessments and further_information_requests
These columns are used to run reports for SLA's within BigQuery where the concept of working days is not possible. This is something that our system had already been doing but the columns and their purpose are now very much out of date as the system had developed over the last 2 years. This commit adds any new working days columns required, removes columns any that are no longer used and bring no value and renaming columns to ensure there's a consistency in the pattern of naming them and making it very clear for data analysts on what they represent.