Update assessor content for existing applications submitted that still have english_language_selt_expired_during_assessment as an FI option.
This is because we are not doing any migration on existing applications to remove this for SELT applications so this is to let the assessors know to just simply not use it.

There is also no risk in assessors using it as it simply just sends the FI to the applicant.