Redirect applicants from "teaching role work history" spoke to "Other work experience in England" spoke if they have not yet started it
This allows a natural flow into the "Other work experience in England" if it is their first time without taking the user to the main application task list. Updates to the normal work history spoke after completing both of these steps will no longer link the two and instead takes the user to the application task list instead.