Generate PrioritisationWorkHistoryCheck records for assessments
If an application has valid work history within the last 12 months that is in England, new prioritisation check records are generated for the assessment. This is only generated if the application form had been generated during the time that the feature was enabled.

If there are new prioritisation checks on the assessment and the assessment previously required preliminary checks, these checks are no longer generated.