Functionality for assessors to accept or reject roles that may be eligibile for prioritisatation checks
Prioritisation checks sit within the pre-assessment stage of the application considerations where LoPs will be sent out for Alberta (Canada) and Hong Kong applicants.

As part of this I have also introduces new columns for the future where a prioritisation decision will be made and whether an application is prioritised.