Introduce new column requires_private_email_for_referee and started_with_private_email_for_referee on application_forms table
The purpose of started_with_private_email_for_referee is to differentiate between applications that began before and after the release of the feature. I see this column as a temporary column primarily used for certain features targetted to existing draft applications after the release which should automatically become irrelevant within 6 months.

The purpose of requires_private_email_for_referee is to ensure the system is aware which applications are submit to the new rules