Fix issue with legacy assessments holding old reference FI reason
Previously we used to have "school_details_cannot_be_verified" as the selected failure reason which prompted assessors to request new reference details. However, this was incorrect and we changed this to "unrecognised_references". Following the latest updates on changes for unrecognised references, we didn't consider legacy applications which would generally be in read only state.

This update fixes this issue to ensure older applications can still be viewed.